src/requirements.txt
python-dotenv==0.21.1
mistralai==1.1.0
psycopg2==2.9.10
pylint==3.3.1